% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{clear_decimal_flags}
\alias{clear_decimal_flags}
\title{Clear sticky decimal flags}
\usage{
clear_decimal_flags()
}
\value{
The previously active sticky flags, invisibly.
}
\description{
Clear sticky decimal flags
}
\examples{
clear_decimal_flags()
}
