% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{get_decimal_context}
\alias{get_decimal_context}
\title{Get the active decimal arithmetic context}
\usage{
get_decimal_context()
}
\value{
A \code{decimal_context} object.
}
\description{
Get the active decimal arithmetic context
}
\examples{
get_decimal_context()
}
