% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal.R
\name{is_snan}
\alias{is_snan}
\title{Identify signaling NaN values}
\usage{
is_snan(x)
}
\arguments{
\item{x}{A decimal-compatible vector.}
}
\value{
A logical vector with the same length as \code{x}.
}
\description{
\code{is_snan()} identifies signaling not-a-number values without performing an
arithmetic operation or raising the \code{invalid_operation} signal.
}
\examples{
is_snan(decimal(c("sNaN", "NaN", "1")))
}
