% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal.R
\name{is_zero}
\alias{is_zero}
\title{Identify decimal zeros}
\usage{
is_zero(x)
}
\arguments{
\item{x}{A decimal-compatible vector.}
}
\value{
A logical vector with the same length as \code{x}.
}
\description{
\code{is_zero()} identifies both positive and negative zero, regardless of the
vector's scale.
}
\examples{
is_zero(decimal(c("0.00", "-0", "1")))
}
