% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{local_decimal_context}
\alias{local_decimal_context}
\title{Install a decimal context for the current scope}
\usage{
local_decimal_context(x, .local_envir = parent.frame())
}
\arguments{
\item{x}{A \code{decimal_context} object or compatible list.}

\item{.local_envir}{Environment whose scope should control restoration.}
}
\value{
\code{x}, invisibly.
}
\description{
Install a decimal context for the current scope
}
\examples{
f <- function() {
  local_decimal_context(decimal_context(precision = 2L, traps = character()))
  decimal("1.234") + decimal("0")
}
f()
}
