% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal.R
\name{NA_decimal_}
\alias{NA_decimal_}
\title{Missing decimal scalar}
\format{
A length-one \code{decimal} vector.
}
\usage{
NA_decimal_
}
\description{
A length-one missing decimal value.
}
\examples{
c(decimal("1"), NA_decimal_)
}
