% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal.R
\name{quantize}
\alias{quantize}
\alias{normalize}
\alias{fma}
\alias{same_quantum}
\title{Decimal-specific operations}
\usage{
quantize(x, quantum)

normalize(x)

fma(x, y, z)

same_quantum(x, y)
}
\arguments{
\item{x, y, z, quantum}{Decimal-compatible vectors.}
}
\value{
A decimal vector, except \code{same_quantum()} which returns a logical
vector.
}
\description{
Decimal-specific operations
}
\examples{
quantize(decimal("1.23456"), decimal("0.01"))
normalize(decimal("1.2300"))
fma(decimal("2"), decimal("3"), decimal("4"))
same_quantum(decimal("1.00"), decimal("2.00"))
}
