% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{set_decimal_context}
\alias{set_decimal_context}
\title{Set the active decimal arithmetic context}
\usage{
set_decimal_context(x)
}
\arguments{
\item{x}{A \code{decimal_context} object or compatible list.}
}
\value{
The previously active \code{decimal_context}, invisibly.
}
\description{
Set the active decimal arithmetic context
}
\examples{
old <- set_decimal_context(decimal_context(precision = 5L))
set_decimal_context(old)
}
