% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{with_decimal_context}
\alias{with_decimal_context}
\title{Use a decimal context within a block}
\usage{
with_decimal_context(x, code)
}
\arguments{
\item{x}{A \code{decimal_context} object or compatible list.}

\item{code}{Code evaluated with \code{x} installed as the active context.}
}
\value{
The result of \code{code}.
}
\description{
Use a decimal context within a block
}
\examples{
with_decimal_context(
  decimal_context(precision = 3L, traps = character()),
  decimal("1.25") + decimal("0")
)
}
