% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal.R
\name{decimal}
\alias{decimal}
\alias{as.decimal}
\alias{is.decimal}
\title{Decimal Class}
\usage{
decimal(x, digits = NULL, decimal_sep = ".", thousand_sep = ",")

as.decimal(x)

is.decimal(x)
}
\description{
`decimal()` converts a valid vector into decimal class;
`is.decimal()` tests whether the object is decimal class;
`as.decimal()` converts suitable vector into decimal class
}
