% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{decimal_flags}
\alias{decimal_flags}
\title{Read sticky decimal flags}
\usage{
decimal_flags()
}
\value{
A character vector of active sticky flags.
}
\description{
Read sticky decimal flags
}
\examples{
decimal_flags()
}
