% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal.R
\name{is_qnan}
\alias{is_qnan}
\title{Identify quiet NaN values}
\usage{
is_qnan(x)
}
\arguments{
\item{x}{A decimal-compatible vector.}
}
\value{
A logical vector with the same length as \code{x}.
}
\description{
\code{is_qnan()} identifies quiet not-a-number values. Signaling NaNs and R
missing values are not quiet NaNs.
}
\examples{
is_qnan(decimal(c("NaN", "sNaN", "1")))
}
